requirements.txt
prysmai>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
